==> notebooks/01_load_and_validate_raw_date.ipynb <==
Descargar data

from pathlib import Path 
import requests

def donwload_one_file_or_raw_data(year: int, month:int) -> Path:
    # Create the directory if it doesn't exist
    output_dir = Path('../data/raw')
    output_dir.mkdir(parents=True, exist_ok=True)
    
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    response = requests.get(URL)

    if response.status_code == 200:
        output_path = output_dir / f'rides_{year}-{month:02d}.parquet'
        output_path.write_bytes(response.content)
        return output_path
    else:
        raise Exception(f"Failed to download file from {URL}")

donwload_one_file_or_raw_data(2024, 1)

import pandas as pd

df = pd.read_parquet('../data/raw/rides_2024-01.parquet')
df.head(20)

Elegimos las columnas relevantes para nuestro probelma

rides = df[['tpep_pickup_datetime', 'PULocationID']]

Las renombramos 

rides.rename(columns=
             {'tpep_pickup_datetime': 'pickup_datetime',
              'PULocationID': 'pickup_location_id'},
                inplace=True)

rides.head(20)

Validamos que el dataset es correcto

rides['pickup_datetime'].describe()

No lo es, ya que hay viajes de 2008 y mayo 2022, por lo que lo corregimos

rides = rides[rides['pickup_datetime'] >= '2024-01-01']
rides = rides[rides['pickup_datetime'] <= '2024-02-01']
rides['pickup_datetime'].describe()

rides.to_parquet('../data/transformed/rides_2024-01.parquet')
