==> notebooks/09_lightgbm_model_with_feature_engineering_pickup_locationID.ipynb <==
exec(open('../src/setup_env.py').read())


import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2024, 10, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    """
    Adds one column with the average rides from
    - 7 days ago
    - 14 days ago
    - 21 days ago
    - 28 days ago
    """
    X['average_rides_last_4_weeks'] = 0.25*(
        X[f'rides_previous_{7*24}_hour'] + \
        X[f'rides_previous_{2*7*24}_hour'] + \
        X[f'rides_previous_{3*7*24}_hour'] + \
        X[f'rides_previous_{4*7*24}_hour']
    )
    return X

from sklearn.preprocessing import FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(
    average_rides_last_4_weeks, validate=False)

add_feature_average_rides_last_4_weeks.fit_transform(X_train)


from sklearn.base import BaseEstimator, TransformerMixin

class TemporalFeaturesEngineer(BaseEstimator, TransformerMixin):

    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        
        X_ = X.copy()
        
        # Generate numeric columns from datetime
        X_["hour"] = X_['pickup_hour'].dt.hour
        X_["day_of_week"] = X_['pickup_hour'].dt.dayofweek
        
        return X_.drop(columns=['pickup_hour'])

add_temporal_features = TemporalFeaturesEngineer()
add_temporal_features.fit_transform(X_train)

def create_nyc_taxi_zones_mapping():
    """
    Crea un diccionario que mapea pickup_location_id a coordenadas (latitud, longitud)
    """
    nyc_zones_coords = {
        1: (40.7648, -73.9808),   # Newark Airport
        2: (40.6173, -74.0065),   # Jamaica Bay
        3: (40.6301, -73.7874),   # Allerton/Pelham Gardens
        4: (40.6818, -73.8371),   # Alphabet City
        5: (40.7307, -73.9866),   # Arden Heights
        6: (40.5564, -74.1756),   # Arrochar/Fort Wadsworth
        7: (40.6326, -74.0793),   # Astoria
        8: (40.7676, -73.9218),   # Astoria Park
        9: (40.7712, -73.9200),   # Auburndale
        10: (40.7388, -73.7716),  # Baisley Park
        11: (40.6957, -73.7596),  # Bath Beach
        12: (40.6016, -74.0156),  # Battery Park
        13: (40.7073, -74.0134),  # Battery Park City
        14: (40.7091, -74.0118),  # Bay Ridge
        15: (40.6220, -74.0330),  # Bay Terrace/Fort Totten
        16: (40.7943, -73.7834),  # Bayside
        17: (40.7681, -73.7639),  # Bedford
        18: (40.6867, -73.9560),  # Bedford-Stuyvesant
        19: (40.6103, -73.9629),  # Belt Parkway
        20: (40.6087, -73.9868),  # Bensonhurst East
        21: (40.6047, -73.9863),  # Bensonhurst West
        22: (40.6063, -74.1084),  # Bloomfield/Emerson Hill
        23: (40.7979, -73.9654),  # Bloomingdale
        24: (40.6883, -73.9819),  # Boerum Hill
        25: (40.6345, -73.9969),  # Borough Park
        26: (40.7179, -73.9938),  # Bowery
        27: (40.5607, -73.9156),  # Breezy Point/Fort Tilden/Riis Beach
        28: (40.5778, -73.9608),  # Brighton Beach
        29: (40.6081, -73.8237),  # Broad Channel
        30: (40.8479, -73.8677),  # Bronx Park
        31: (40.8404, -73.8467),  # Bronxdale
        32: (40.6966, -73.9954),  # Brooklyn Heights
        33: (40.7003, -73.9784),  # Brooklyn Navy Yard
        34: (40.6533, -73.9084),  # Brownsville
        35: (40.6942, -73.9180),  # Bushwick North
        36: (40.6835, -73.9248),  # Bushwick South
        37: (40.6893, -73.7367),  # Cambria Heights
        38: (40.6467, -73.9056),  # Canarsie
        39: (40.6750, -73.9969),  # Carroll Gardens
        40: (40.8068, -73.9478),  # Central Harlem
        41: (40.8138, -73.9440),  # Central Harlem North
        42: (40.7829, -73.9654),  # Central Park
        43: (40.7829, -73.9654),  # Central Park (duplicate ID fixed)
        44: (40.7161, -73.9961),  # Chinatown
        45: (40.8484, -73.7864),  # City Island
        46: (40.8404, -73.9162),  # Claremont/Bathgate
        47: (40.7602, -73.9936),  # Clinton East
        48: (40.6881, -73.9688),  # Clinton Hill
        49: (40.7602, -73.9936),  # Clinton West
        50: (40.8745, -73.8298),  # Co-Op City
        51: (40.6874, -73.9968),  # Cobble Hill
        52: (40.7850, -73.8447),  # College Point
        53: (40.6775, -74.0038),  # Columbia Street
        54: (40.5755, -73.9707),  # Coney Island
        55: (40.7514, -73.8631),  # Corona
        56: (40.7631, -73.7510),  # Douglaston
        57: (40.6929, -73.9851),  # Downtown Brooklyn/MetroTech
        58: (40.7018, -73.9890),  # DUMBO/Vinegar Hill
        59: (40.6199, -74.0277),  # Dyker Heights
        60: (40.7474, -73.9963),  # East Chelsea
        61: (40.8217, -73.9255),  # East Concourse/Concourse Village
        62: (40.7602, -73.8789),  # East Elmhurst
        63: (40.6504, -73.9244),  # East Flatbush/Farragut
        64: (40.6356, -73.9253),  # East Flatbush/Remsen Village
        65: (40.7957, -73.9389),  # East Harlem North
        66: (40.7898, -73.9422),  # East Harlem South
        67: (40.6663, -73.8827),  # East New York
        68: (40.6590, -73.8973),  # East New York/Pennsylvania Avenue
        69: (40.8406, -73.8906),  # East Tremont
        70: (40.7281, -73.9840),  # East Village
        71: (40.7155, -73.9396),  # East Williamsburg
        72: (40.8871, -73.8274),  # Eastchester
        73: (40.7388, -73.8789),  # Elmhurst
        74: (40.7420, -73.8795),  # Elmhurst/Maspeth
        75: (40.5387, -74.2009),  # Eltingville/Annadale/Prince's Bay
        76: (40.5913, -74.1503),  # Emerson Hill
        77: (40.6047, -73.7631),  # Far Rockaway
        78: (40.7074, -74.0113),  # Financial District North
        79: (40.7063, -74.0116),  # Financial District South
        80: (40.6337, -73.9594),  # Flatbush/Ditmas Park
        81: (40.7410, -73.9896),  # Flatiron
        82: (40.6186, -73.9393),  # Flatlands
        83: (40.7673, -73.8333),  # Flushing
        84: (40.7498, -73.8428),  # Flushing Meadows-Corona Park
        85: (40.8619, -73.9016),  # Fordham South
        86: (40.7184, -73.8441),  # Forest Hills
        87: (40.6919, -73.9740),  # Fort Greene
        88: (40.7347, -73.7936),  # Fresh Meadows
        89: (40.5707, -74.1866),  # Freshkills Park
        90: (40.7548, -73.9933),  # Garment District
        91: (40.7475, -73.7118),  # Glen Oaks
        92: (40.6736, -73.9847),  # Gowanus
        93: (40.7368, -73.9848),  # Gramercy
        94: (40.5912, -73.9772),  # Gravesend
        95: (40.5563, -74.1571),  # Great Kills
        96: (40.7306, -73.9496),  # Greenpoint
        97: (40.7335, -74.0027),  # Greenwich Village
        98: (40.6155, -74.0753),  # Grymes Hill/Clifton
        99: (40.8230, -73.9494),  # Hamilton Heights
        100: (40.5899, -73.8047), # Hammels/Arverne
        101: (40.5968, -74.1399), # Heartland Village/Todt Hill
        102: (40.8392, -73.9276), # Highbridge
        103: (40.7122, -73.7624), # Hollis
        104: (40.5949, -73.9735), # Homecrest
        105: (40.6659, -73.8304), # Howard Beach
        106: (40.8131, -73.8823), # Hunts Point
        107: (40.8681, -73.9211), # Inwood
        108: (40.8681, -73.9273), # Inwood Hill Park
        109: (40.7022, -73.7978), # Jamaica
        110: (40.7165, -73.7837), # Jamaica Estates
        111: (40.6464, -73.9728), # Kensington
        112: (40.7077, -73.8315), # Kew Gardens
        113: (40.7280, -73.8248), # Kew Gardens Hills
        114: (40.7424, -73.9815), # Kips Bay
        115: (40.7464, -73.9857), # Korea Town
        116: (40.8492, -73.8635), # Laconia
        117: (40.7669, -73.8740), # LaGuardia Airport
        118: (40.6680, -73.7481), # Laurelton
        119: (40.7738, -73.9573), # Lenox Hill East
        120: (40.7653, -73.9629), # Lenox Hill West
        121: (40.7813, -73.9807), # Lincoln Square East
        122: (40.7813, -73.9886), # Lincoln Square West
        123: (40.7214, -73.9968), # Little Italy/NoLiTa
        124: (40.7447, -73.9485), # Long Island City/Hunters Point
        125: (40.7506, -73.9380), # Long Island City/Queens Plaza
        126: (40.8162, -73.8901), # Longwood
        127: (40.7155, -73.9851), # Lower East Side
        128: (40.7267, -73.9108), # Maspeth
        129: (40.7390, -74.0089), # Meatpacking/West Village West
        130: (40.8267, -73.9137), # Melrose South
        131: (40.7184, -73.8795), # Middle Village
        132: (40.7505, -73.9934), # Midtown Center
        133: (40.7614, -73.9776), # Midtown East
        134: (40.7590, -73.9845), # Midtown North
        135: (40.7505, -73.9934), # Midtown South
        136: (40.6177, -73.9654), # Midwood
        137: (40.8073, -73.9624), # Morningside Heights
        138: (40.8259, -73.9063), # Morrisania
        139: (40.8044, -73.9249), # Mott Haven/Port Morris
        140: (40.8462, -73.9179), # Mount Hope
        141: (40.7472, -73.9764), # Murray Hill
        142: (40.7605, -73.8917), # Murray Hill-Queens
        143: (40.5806, -74.1177), # New Dorp/Midland Beach
        144: (40.5939, -74.1540), # New Springville/Bloomfield/Travis
        145: (40.7507, -73.8673), # North Corona
        146: (40.8811, -73.8778), # Norwood
        147: (40.7304, -73.7615), # Oakland Gardens
        148: (40.5648, -74.1644), # Oakwood
        149: (40.7644, -73.9300), # Old Astoria
        150: (40.6793, -73.8533), # Ozone Park
        151: (40.6728, -73.9732), # Park Slope
        152: (40.8338, -73.8584), # Parkchester
        153: (40.8654, -73.8078), # Pelham Bay
        154: (40.8654, -73.8078), # Pelham Bay Park
        155: (40.8572, -73.8677), # Pelham Parkway
        156: (40.7507, -73.9934), # Penn Station/Madison Sq West
        157: (40.6344, -74.1394), # Port Richmond
        158: (40.6608, -73.9608), # Prospect-Lefferts Gardens
        159: (40.6775, -73.9680), # Prospect Heights
        160: (40.6621, -73.9692), # Prospect Park
        161: (40.7505, -73.9934), # Queens Village
        162: (40.7614, -73.9776), # Queensboro Hill
        163: (40.7590, -73.9845), # Queensbridge/Ravenswood
        164: (40.7505, -73.9934), # Randalls Island
        165: (40.6751, -74.0064), # Red Hook
        166: (40.7264, -73.8535), # Rego Park
        167: (40.7008, -73.8319), # Richmond Hill
        168: (40.7063, -73.9064), # Ridgewood
        169: (40.7930, -73.8737), # Rikers Island
        170: (40.5810, -73.8370), # Rockaway Park
        171: (40.7615, -73.9492), # Roosevelt Island
        172: (40.5378, -74.2067), # Rossville/Woodrow
        173: (40.6998, -73.7635), # Saint Albans
        174: (40.6436, -74.0778), # Saint George/New Brighton
        175: (40.7458, -73.9068), # Saint Michaels Cemetery/Woodside
        176: (40.8271, -73.8226), # Schuylerville/Throgs Neck
        177: (40.7069, -74.0033), # Seaport
        178: (40.5882, -73.9441), # Sheepshead Bay
        179: (40.7239, -74.0027), # SoHo
        180: (40.8154, -73.8706), # Soundview/Bruckner
        181: (40.8271, -73.8513), # Soundview/Castle Hill
        182: (40.5918, -74.1043), # South Beach/Dongan Hills
        183: (40.6890, -73.7909), # South Jamaica
        184: (40.6687, -73.8111), # South Ozone Park
        185: (40.7127, -73.9574), # South Williamsburg
        186: (40.6657, -73.7607), # Springfield Gardens North
        187: (40.6506, -73.7641), # Springfield Gardens South
        188: (40.6274, -74.0776), # Stapleton
        189: (40.6487, -73.8863), # Starrett City
        190: (40.7717, -73.9032), # Steinway
        191: (40.6825, -73.9384), # Stuyvesant Heights
        192: (40.7357, -73.9794), # Stuyvesant Town/PCV
        193: (40.6553, -74.0093), # Sunset Park East
        194: (40.6420, -74.0130), # Sunset Park West
        195: (40.7576, -73.9608), # Sutton Place/Turtle Bay East
        196: (40.7435, -73.9240), # Sunnyside
        197: (40.7576, -73.9608), # Sutton Place/Turtle Bay East
        198: (40.7589, -73.9851), # Times Sq/Theatre District
        199: (40.5964, -74.1434), # Todt Hill/Emerson Hill
        200: (40.7195, -74.0137), # Tribeca/Civic Center
        201: (40.7129, -73.9851), # Two Bridges/Seward Park
        202: (40.7359, -73.9906), # Union Sq
        203: (40.8512, -73.9162), # University Heights/Morris Heights
        204: (40.7851, -73.9534), # Upper East Side North
        205: (40.7737, -73.9566), # Upper East Side South
        206: (40.7870, -73.9754), # Upper West Side North
        207: (40.7937, -73.9684), # Upper West Side South
        208: (40.8896, -73.8965), # Van Cortlandt Village
        209: (40.8896, -73.8965), # Van Cortlandt Park
        210: (40.8448, -73.8648), # Van Nest/Morris Park
        211: (40.8505, -73.9344), # Washington Heights North
        212: (40.8418, -73.9393), # Washington Heights South
        213: (40.6240, -74.1424), # West Brighton
        214: (40.7578, -74.0020), # West Chelsea/Hudson Yards
        215: (40.7335, -74.0088), # West Village
        216: (40.8407, -73.8428), # Westchester Square
        217: (40.6063, -74.1234), # Westerleigh
        218: (40.7944, -73.8075), # Whitestone
        219: (40.7544, -73.8606), # Willets Point
        220: (40.8796, -73.8651), # Williamsbridge/Olinville
        221: (40.7056, -73.9444), # Williamsburg (North Side)
        222: (40.7056, -73.9539), # Williamsburg (South Side)
        223: (40.6572, -73.9759), # Windsor Terrace
        224: (40.6896, -73.8581), # Woodhaven
        225: (40.8976, -73.8612), # Woodlawn/Wakefield
        226: (40.7458, -73.9068), # Woodside
        227: (40.7115, -74.0134), # World Trade Center
        228: (40.7834, -73.9476), # Yorkville East
        229: (40.7834, -73.9584), # Yorkville West
        230: (39.1638, -75.5264), # NV (Nevada)
        231: (None, None),        # NA (Not Available)
        232: (40.6413, -74.0838), # Rockaways
        233: (40.5920, -73.8154), # Soundview/Castle Hill
        234: (40.8271, -73.8513), # Union Sq
        235: (40.7359, -73.9906), # Van Cortlandt Village
        236: (40.8896, -73.8965), # Williamsburg (North Side)
        237: (40.7589, -73.9851), # Times Sq/Theatre District
        238: (40.6413, -74.0838), # Near JFK Airport
        239: (40.6892, -73.7781), # Bronx Zoo
        240: (40.8506, -73.8772), # Central Harlem
        241: (40.8068, -73.9478), # Central Harlem North
        242: (40.8138, -73.9440), # Inwood
        243: (40.8681, -73.9211), # Washington Heights North
        244: (40.8505, -73.9344), # Washington Heights South
        245: (40.8418, -73.9393), # Inwood Hill Park
        246: (40.8681, -73.9273), # Marble Hill
        247: (40.8764, -73.9107), # Bronx Park
        248: (40.8479, -73.8677), # Van Cortlandt Park
        249: (40.8896, -73.8965), # Pelham Bay Park
        250: (40.8654, -73.8078), # Co-Op City
        251: (40.8745, -73.8298), # Westchester Square
        252: (40.8407, -73.8428), # City Island
        253: (40.8484, -73.7864), # Parkchester
        254: (40.8338, -73.8584), # Clason Point
        255: (40.8154, -73.8706), # Hunts Point
        256: (40.8131, -73.8823), # Longwood
        257: (40.8162, -73.8901), # Melrose South
        258: (40.8267, -73.9137), # Mott Haven/Port Morris
        259: (40.8044, -73.9249), # Morrisania
        260: (40.8259, -73.9063), # Mount Hope
        261: (40.8462, -73.9179), # University Heights/Morris Heights
        262: (40.8512, -73.9162), # Fordham South
        263: (40.8619, -73.9016), # East Tremont
        264: (40.8406, -73.8906), # Belmont
        265: (40.8554, -73.8804)  # West Farms/Bronx River
    }
    return nyc_zones_coords

from sklearn.base import BaseEstimator, TransformerMixin

class CoordinatesEngineer(BaseEstimator, TransformerMixin):

    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        
        X_ = X.copy()
        coords_mapping = create_nyc_taxi_zones_mapping()
        
        def safe_get_lat(location_id):
            try:
                coords = coords_mapping.get(location_id, (None, None))
                if coords is None or coords == (None, None):
                    return None
                return coords[0]
            except:
                return None
        
        def safe_get_lon(location_id):
            try:
                coords = coords_mapping.get(location_id, (None, None))
                if coords is None or coords == (None, None):
                    return None
                return coords[1]
            except:
                return None
        
        # Generate coordinate columns from pickup_location_id
        X_['pickup_lat'] = X_['pickup_location_id'].map(safe_get_lat)
        X_['pickup_lon'] = X_['pickup_location_id'].map(safe_get_lon)
        
        return X_.drop(columns=['pickup_location_id'])

add_temporal_coordinates = CoordinatesEngineer()
add_temporal_coordinates.fit_transform(X_train)

import lightgbm as lgb

from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_temporal_coordinates,
    add_feature_average_rides_last_4_weeks,
    add_temporal_features,
    lgb.LGBMRegressor()
)
pipeline.fit(X_train, y_train)
     

predictions = pipeline.predict(X_test)

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
