==> notebooks/14_inference_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

from datetime import datetime
import pandas as pd


current_date = pd.to_datetime(datetime.utcnow()).floor('H')
print(f'{current_date=}')

from src.inference import load_batch_of_features_from_store_v2

features = load_batch_of_features_from_store_v2(current_date)

from src.inference import load_model_from_registry, get_model_predictions


model = load_model_from_registry()
predictions = get_model_predictions(model, features)

# Ajustar el DataFrame para el feature group de predicciones
predictions['pickup_ts'] = current_date  # Usar pickup_ts en lugar de pickup_hour
predictions.rename(columns={'predicted_demand': 'predictions'}, inplace=True)  # Renombrar para consistencia
predictions

from src.feature_store_api import get_or_create_feature_group
import src.config as config

# Usar el feature group correcto para predicciones
feature_group = get_or_create_feature_group(config.FEATURE_GROUP_PREDICTIONS_METADATA)

feature_group.insert(predictions, write_options={"wait_for_job": False})
